LayoutTests/platform/mac/accessibility/aria-tree-expected.txt

Animals
Birds
Cats
Siamese
Tabby
This tests that aria trees and tree items are converted into AXOutlines and AXRows, with all the right attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tree.role is 'AXRole: AXOutline'
PASS canSetRows is true
PASS tree.childrenCount is 5
PASS treeitem1.role is 'AXRole: AXRow'
PASS treeitem1.subrole is 'AXSubrole: AXOutlineRow'
PASS treeitem1.isExpanded is true
PASS treeitem1.hierarchicalLevel is 1
PASS canSetDisclosing is true
PASS treeitem2.role is 'AXRole: AXRow'
PASS treeitem2.subrole is 'AXSubrole: AXOutlineRow'
PASS treeitem2.isExpanded is false
PASS treeitem2.hierarchicalLevel is 2
PASS treeitem2.disclosedByRow().isEqual(treeitem1) is true
PASS canSetDisclosing is false
PASS treeitem3.stringValue is 'AXValue: Birds'
PASS canSetDisclosing is false
PASS treeitem3.role is 'AXRole: AXRow'
PASS treeitem3.subrole is 'AXSubrole: AXOutlineRow'
PASS treeitem3.isExpanded is true
PASS treeitem3.isSelected is false
PASS treeitem3.isSelected is true
PASS selectedRow.isEqual(treeitem3) is true
PASS successfullyParsed is true

TEST COMPLETE
